Update Settings/SettingsController to use `null` instead of empty string.
When use empty string for no value and check with no value to use default, it is not working. Use `null` instead for some config.
Example: `$this->configDb['rdbadmin_UserLoginMaxFail'] ?? 10` will always return empty string when this config value is empty string and cause argument type error in PHP.
If the no value is `null` then this works fine on everywhere.